+++ BudgetApp.Web/Controllers/ExpensesController.cs
@@ -7,9 +7,11 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using BudgetApp.Web.Data;
 using BudgetApp.Web.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace BudgetApp.Web.Controllers
 {
+    [Authorize]
     public class ExpensesController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -19,42 +21,30 @@ namespace BudgetApp.Web.Controllers
             _context = context;
         }
 
-        // GET: Expenses
         public async Task<IActionResult> Index()
         {
             var applicationDbContext = _context.Expenses.Include(e => e.Category);
             return View(await applicationDbContext.ToListAsync());
         }
 
-        // GET: Expenses/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var expense = await _context.Expenses
                 .Include(e => e.Category)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (expense == null)
-            {
-                return NotFound();
-            }
+            if (expense == null) return NotFound();
 
             return View(expense);
         }
 
-        // GET: Expenses/Create
         public IActionResult Create()
         {
             ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Name");
             return View();
         }
 
-        // POST: Expenses/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Title,Amount,Date,CategoryId")] Expense expense)
@@ -69,34 +59,22 @@ namespace BudgetApp.Web.Controllers
             return View(expense);
         }
 
-        // GET: Expenses/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var expense = await _context.Expenses.FindAsync(id);
-            if (expense == null)
-            {
-                return NotFound();
-            }
+            if (expense == null) return NotFound();
+
             ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Name", expense.CategoryId);
             return View(expense);
         }
 
-        // POST: Expenses/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Amount,Date,CategoryId")] Expense expense)
         {
-            if (id != expense.Id)
-            {
-                return NotFound();
-            }
+            if (id != expense.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -107,14 +85,8 @@ namespace BudgetApp.Web.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ExpenseExists(expense.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!ExpenseExists(expense.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -122,35 +94,24 @@ namespace BudgetApp.Web.Controllers
             return View(expense);
         }
 
-        // GET: Expenses/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var expense = await _context.Expenses
                 .Include(e => e.Category)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (expense == null)
-            {
-                return NotFound();
-            }
+            if (expense == null) return NotFound();
 
             return View(expense);
         }
 
-        // POST: Expenses/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var expense = await _context.Expenses.FindAsync(id);
-            if (expense != null)
-            {
-                _context.Expenses.Remove(expense);
-            }
+            if (expense != null) _context.Expenses.Remove(expense);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
